refactor(hickey): extract TabStrip components to TabStrip.tsx
Addresses Hickey F3: App.tsx had grown to four concerns — multi-host
shell (App), tab strip (TabStrip/TabChip/AddHostForm), per-host body
(HostView), and per-host helpers (Header/CpuStrip/FilterBar/...). Pull
TabStrip + TabChip + AddHostForm into their own file so a tab-UI
redesign and a host-view redesign no longer share a file.

TabStrip.tsx owns the chip's connection-cell subscription (display
volatility) but explicitly does NOT own per-host socket disposal — that
lives in MultiHostApp's host-removal effect (transport-lifetime
volatility), per Lowy L4.
